Lots of MIDI editor peripheral stuff
 * Saving, loading, exporting, importing MIDI (well not exporting yet, still have to wire that last one up) including full backend implementation for saving MIDI compositions in the database.
  * Lots of pieces for this including a few custom modals, initialization logic, de/serialization, etc.
  * Re-initialization function for MIDI Editor instance so we can load a new state into an existing instance
  * Set up `react-query` with helper functions adding the query client, provider, etc. for React components
 * Probably the whole MIDI editor toolbar with play/pause, BPM select, beat snap interval select, and other stuff. Now that I think of it, I haven't committed anything in a long time I think

+++ engine/common/src/lib.rs
@@ -18,8 +18,8 @@ pub use crate::init::*;
 #[derive(Serialize, Deserialize)]
 pub struct RawNoteData {
     pub line_ix: usize,
-    pub start_beat: f32,
-    pub width: f32,
+    pub start_beat: f64,
+    pub width: f64,
 }
 
 // Transmuted `rand_pcg::Pcg32::new(0xcafef00dd15ea5e5, 0xa02bdbf7bb3c0a7)` since it's not const atm
